backend/requirements.txt
fastapi
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-docx>=1.1.0
python-pptx>=0.6.21
PyPDF2>=3.0.1
python-dotenv>=1.0.0
pandas>=2.1.0
Markdown>=3.5.0
chardet>=5.2.0
openpyxl>=3.1.2
reportlab>=4.0.4
requests>=2.31.0
numpy>=1.26.0
openai
python-http-client>=3.3.7
email-validator>=2.1.0
google-genai
aiohttp
supabase
pypandoc